Make posix-compliant
Posix-compiant shells, such as dash,
which is the default `/bin/sh` on Debian,
don't support escape characters in the form
of `echo -e \e[xxxx $TEXT`.

As a result, dash outputs this on `./tuxi --help`:
```
-e \e[1;32mUsage:\e[0m tuxi \e[1;35m<your question>\e[0m
       tuxi \e[1;36m<OPTIONS>\e[0m \e[1;35m<your question>\e[0m
```

Dash, and other non-bash shells support coloring in the form
of `printf '%b' "\033[xxxx $TEXT"`.

This patch changes the coloring.