Add first pass of Tuxi man page
This needs various sections filled out, but I've added some content
already. Do not release this to the main branch; it's very much
incomplete.

I've come up with a short description, but feel free to use something
different. You don't want this to be too long; it should be concise, -
not end in a period ('.'), and if it starts with an alphabetical
character, it should be lowercase; that's at least the Debian and
Ubuntu standard for package short descriptions, which are usually
synonymous with those found in man(1) pages.

I've included standard formatting to the man page, appropriate for
those using colors via the likes of `LESS_TERMCAP_*`.

I apologise if I got anything wrong, as I just paraphrased and
sometimes copied what was in the usage output. See the next commit for
that stuff.

Add to EXAMPLES, as it's referenced in one of the flags. The usage for
Tuxi is a little strange, so it'd be good to add some examples.

For NOTES, you could add, for example, how things work, nuances to
Tuxi, caveats, etc.

You may want to add a SEE ALSO section at the very bottom:

```
.SH "SEE ALSO"
\fIlike-this\fR(1), \fIor-this\fR(4)
```

Something like that.

If there are any other files to be added (files created or perhaps even
those used by Tuxi), you're probably best to add them to the FILES
section.